requirements.txt
python-dotenv>=1.0.0
selenium>=4.0.0
webdriver-manager>=4.0.0
pygetwindow>=0.0.9

